requirements.txt
requests
python-dotenv
crewai
fastapi
crewai[tools]
PyGithub
google.generativeai
itsdangerous
langchain-core
litellm
crewai[google-genai]
